+++ example/locales/nl-NL.js
@@ -1,0 +1,3 @@
+export default {
+  name: '你好，{name}',
+};