+++ packages/plugin-helmet/fixtures/ssr/.umirc.ts
@@ -1,0 +1,14 @@
+
+export default {
+  ssr: {},
+  history: {
+    type: 'memory',
+    options: {
+      initialEntries: ['/'],
+    },
+  },
+  mountElementId: '',
+  routes: [
+    { path: '/', component: 'index' },
+  ],
+}